feat: add upcoming appointments service
